Improve docker setup.

+++ mp.sh
@@ -2,13 +2,17 @@ _mp() {
     docker-compose -f .docker-compose.yml run --rm --service-ports mailpile "$@"
 }
 
-mp-setup() {
-    docker volume rm mailpile_data
-    _mp setup
-    _mp --www=0.0.0.0:33411
-}
-
 mp() {
+    if [ "$1" = "setup" ]; then
+        shift
+        docker volume rm mailpile_data
+        if [ "$?" != 0 ]; then
+            return "$?"
+        fi
+        _mp setup
+        _mp --www=0.0.0.0:33411
+    fi
+
     echo "http://127.0.0.1:33411"
     _mp "$@"
 }
